Added Tag component, moving tags in cards below title
