+++ src/components/Card/Card.js
@@ -11,36 +11,6 @@ import {
   tag,
 } from "./Card.module.scss";
 
-const getTypeString = (type, link) => {
-  if (type === "default" && link.indexOf("blog") !== -1) {
-    return "blog";
-  }
-  else if(type === "default" && link.indexOf("blog") === -1) {
-    return "review";
-  }
-  else if (type === "component") {
-    if(link.indexOf("3do") !== -1) {
-      return "3do";
-    }
-    else if (link.indexOf("bm") !== -1) {
-      return "bm";
-    }
-    else if (link.indexOf("fme") !== -1) {
-      return "fme";
-    }
-    else if (link.indexOf("voc") !== -1) {
-      return "voc";
-    }
-    else if (link.indexOf("patch") !== -1) {
-      return "patch";
-    }
-    else if (link.indexOf("wax") !== -1) {
-      return "wax";
-    }
-  }
-  return "mission";
-}
-
 const Card = ({
   title,
   coverURL,
@@ -54,8 +24,6 @@ const Card = ({
   if (type === "mission") typeClass = mission;
   else if (type === "component") typeClass = component;
 
-  const tagText = getTypeString(type, link);
-
   return (
     <Link
       to={link}
@@ -65,7 +33,6 @@ const Card = ({
         <div className={cover} style={{ backgroundImage: `url(${coverURL})` }}></div>
       )}
       <article>
-        <span className={tag}>{tagText}</span>
         <h1>
           <Aurebesh text={title} />
           {title}
