+++ src/components/Card/Card.module.scss
@@ -38,6 +38,7 @@
 
     h1 {
         font-size: 2.4rem;
+        line-height: 1.3;
         margin-bottom: 0;
         margin-top: 0;
         transition: all 0.1s ease-in-out;
@@ -46,6 +47,11 @@
             font-size: 1.2rem;
             color: rgba(255,255,255,0.3);
         }
+    }
+
+    :global(.secondary) {
+        display: block;
+        margin-top: 4px;
     }
 
     &:hover {
