+++ src/components/ReviewCard/ReviewCard.js
@@ -1,6 +1,7 @@
 import React from "react";
 import Card from "../Card";
 import Rating from "../Rating";
+import Tag from "../Tag";
 import { authorList } from "./ReviewCard.module.scss";
 
 const ReviewCard = ({ review, orientation }) => {
@@ -9,7 +10,11 @@ const ReviewCard = ({ review, orientation }) => {
 
   return (
     <Card link={`/missions/${mission.slug}`} coverURL={coverURL} title={mission.title} orientation={orientation} type="mission">
-      <small className={authorList}>Reviewed by {review.frontmatter.authors.join(", ")}</small>
+      <div className="secondary">
+        <Tag type="mission" text="Review" />
+        <small className={authorList}>Reviewed by {review.frontmatter.authors.join(", ")}</small>
+      </div>
+      
       <p>{review.excerpt}</p>
       <Rating score={review.frontmatter.rating} />
     </Card>
