+++ src/components/ReviewCard/ReviewCard.module.scss
@@ -3,8 +3,6 @@
 .authorList {
     font-size: 1.4rem;
     color: rgba(255,255,255,0.6);
-    display: block;
-    margin-top: 8px;
 }
 
 .ecClass {
